+++ datanymizer_engine/src/transformers/token/mod.rs
@@ -1,5 +1,3 @@
-use rand::distributions::{Distribution, Uniform};
-
 mod base64;
 pub use base64::Base64TokenTransformer;
 
@@ -8,44 +6,3 @@ pub use base64url::Base64UrlTokenTransformer;
 
 mod hex;
 pub use hex::HexTokenTransformer;
-
-fn rnd_chars(len: usize, src: &[char]) -> String {
-    let rng = rand::thread_rng();
-    let distribution = Uniform::<usize>::from(0..src.len());
-    distribution
-        .sample_iter(rng)
-        .take(len)
-        .map(|i| src[i])
-        .collect::<String>()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn same_char() {
-        let chars = vec!['a'];
-
-        assert_eq!(rnd_chars(0, &chars), "");
-        assert_eq!(rnd_chars(1, &chars), "a");
-        assert_eq!(rnd_chars(4, &chars), "aaaa");
-    }
-
-    #[test]
-    fn different_chars() {
-        let chars = vec!['a', 'b', 'c'];
-
-        assert_eq!(rnd_chars(0, &chars), "");
-
-        let s = rnd_chars(1, &chars);
-        for ch in s.chars() {
-            assert!(chars.contains(&ch));
-        }
-
-        let s = rnd_chars(5, &chars);
-        for ch in s.chars() {
-            assert!(chars.contains(&ch));
-        }
-    }
-}